+++ bookwyrm/templates/feed/direct_messages.html
@@ -1,5 +1,5 @@
-{% extends 'layout.html' %}
-{% block content %}
+{% extends 'feed/feed_layout.html' %}
+{% block panel %}
 
 <div class="block">
     <h1 class="title">Direct Messages</h1>
