+++ bookwyrm/templates/feed/feed.html
@@ -1,0 +1,39 @@
+{% extends 'feed/feed_layout.html' %}
+{% load bookwyrm_tags %}
+{% block panel %}
+
+<h1 class="title">{{ tab | title }} Timeline</h1>
+<div class="tabs">
+    <ul>
+        <li class="{% if tab == 'home' %}is-active{% endif %}">
+            <a href="/#feed">Home</a>
+        </li>
+        <li class="{% if tab == 'local' %}is-active{% endif %}">
+            <a href="/local#feed">Local</a>
+        </li>
+        <li class="{% if tab == 'federated' %}is-active{% endif %}">
+            <a href="/federated#feed">Federated</a>
+        </li>
+    </ul>
+</div>
+
+{# announcements and system messages #}
+{% if not goal and tab == 'home' %}
+{% now 'Y' as year %}
+<section class="block hidden" aria-title="Announcements" data-hide="hide-{{ year }}-reading-goal">
+    {% include 'snippets/goal_card.html' with year=year %}
+    <hr>
+</section>
+{% endif %}
+
+{# activity feed #}
+{% if not activities %}
+<p>There aren't any activities right now! Try following a user to get started</p>
+{% endif %}
+{% for activity in activities %}
+<div class="block">
+{% include 'snippets/status.html' with status=activity %}
+</div>
+{% endfor %}
+
+{% endblock %}
