+++ bookwyrm/templates/feed/status.html
@@ -1,0 +1,13 @@
+{% extends 'feed/feed_layout.html' %}
+{% block panel %}
+<header class="block">
+    <a href="/#feed" class="button" data-back>
+        <span class="icon icon-arrow-left" aira-hidden="true"></span>
+        <span>Back</span>
+    </a>
+</header>
+
+{% include 'feed/thread.html' with status=status depth=0 max_depth=6 is_root=True direction=0 %}
+
+{% endblock %}
+
