+++ bookwyrm/templates/feed/thread.html
