--- bookwyrm/templates/status.html
@@ -1,9 +1,0 @@
-{% extends 'layout.html' %}
-{% block content %}
-
-<div class="block">
-    {% include 'snippets/thread.html' with status=status depth=0 max_depth=6 is_root=True direction=0 %}
-</div>
-
-{% endblock %}
-
