+++ bookwyrm/views/__init__.py
@@ -4,15 +4,15 @@ from .author import Author, EditAuthor
 from .block import Block, unblock
 from .books import Book, EditBook, Editions
 from .books import upload_cover, add_description, switch_edition, resolve_book
-from .direct_message import DirectMessage
 from .error import not_found_page, server_error_page
+from .feed import DirectMessage, Feed, Replies, Status
 from .follow import follow, unfollow
 from .follow import accept_follow_request, delete_follow_request, handle_accept
 from .goal import Goal
 from .import_data import Import, ImportStatus
 from .interaction import Favorite, Unfavorite, Boost, Unboost
 from .invite import ManageInvites, Invite
-from .landing import About, Home, Feed, Discover
+from .landing import About, Home, Discover
 from .notifications import Notifications
 from .outbox import Outbox
 from .reading import edit_readthrough, create_readthrough, delete_readthrough
@@ -24,6 +24,6 @@ from .search import Search
 from .shelf import Shelf
 from .shelf import user_shelves_page, create_shelf, delete_shelf
 from .shelf import shelve, unshelve
-from .status import Status, Replies, CreateStatus, DeleteStatus
+from .status import CreateStatus, DeleteStatus
 from .updates import Updates
 from .user import User, EditUser, Followers, Following
